Cleaned up some int -> NSUInteger warnings fixed some spelling errors

+++ FlyingSwallow/Controller/HelpViewController.m
@@ -13,7 +13,7 @@
     NSMutableArray *pageViewArray;
     NSMutableArray *pageTitleArray;
     
-    int pageCount;
+    NSUInteger pageCount;
 }
 
 @end
